PATCH: fix #3, by updating constructor with a default parameter giving information about the size / part of the sprite used

+++ src/object.hh
@@ -13,23 +13,24 @@ class Object
     public :
         Object(float pos_x,
                 float pos_y,
-                float len_x,
-                float len_y);
+                int len_x,
+                int len_y);
 
-        Object(sf::Vector2f pos, sf::Vector2f len);
+        Object(sf::Vector2f pos,
+                sf::Vector2i len);
 
         // Getters
         /// {
         sf::Vector2f pos_get(void);
-        sf::Vector2f len_get(void);
+        sf::Vector2i len_get(void);
 
         float pos_x_get(void);
         float pos_y_get(void);
-        float len_y_get(void);
-        float len_x_get(void);
+        int len_y_get(void);
+        int len_x_get(void);
         /// }
 
     protected :
         sf::Vector2f pos_;
-        sf::Vector2f len_;
+        sf::Vector2i len_;
 };
